refactor: make custom-rpc lp use the cf-rpc-apis lp trait
