+++ api/bin/chainflip-lp-api/Cargo.toml
@@ -45,6 +45,7 @@ thiserror = { workspace = true, default-features = true }
 # Local
 chainflip-api = { workspace = true }
 cf-utilities = { workspace = true, default-features = true }
+cf-rpc-apis = { workspace = true }
 
 [build-dependencies]
 substrate-build-script-utils = { workspace = true }
