+++ engine/src/state_chain_observer/client/extrinsic_api/signed.rs
@@ -26,7 +26,8 @@ use super::{
 
 use anyhow::Result;
 use async_trait::async_trait;
-use cf_node_client::{signer, WaitFor, WaitForResult};
+use cf_node_client::{signer, WaitForResult};
+use cf_primitives::WaitFor;
 use cf_utilities::task_scope::{task_scope, Scope, ScopedJoinHandle, UnwrapOrCancel};
 use futures::StreamExt;
 use futures_util::FutureExt;
