+++ state-chain/custom-rpc/src/lp.rs
@@ -17,7 +17,6 @@
 use crate::{
 	backend::{CustomRpcBackend, NotificationBehaviour},
 	order_fills,
-	order_fills::OrderFills,
 	pool_client::SignedPoolClient,
 	CfApiError, RpcResult, StorageQueryApi,
 };
@@ -33,23 +32,25 @@ use cf_node_client::{
 	events_decoder::{DynamicEventError, DynamicEvents},
 	extract_from_first_matching_event,
 	subxt_state_chain_config::cf_static_runtime,
-	ApiWaitForResult, ExtrinsicData, WaitFor, WaitForDynamicResult,
+	ExtrinsicData, WaitForDynamicResult,
 };
 use cf_primitives::{
 	chains::{assets::any::AssetMap, Arbitrum, Bitcoin, Ethereum, Polkadot, Solana},
-	Asset, BasisPoints, BlockNumber, DcaParameters, EgressId, ForeignChain, Price,
+	ApiWaitForResult, Asset, BasisPoints, BlockNumber, DcaParameters, EgressId, ForeignChain,
+	Price, WaitFor,
 };
+use cf_rpc_apis::lp::LpRpcApiServer;
 use cf_rpc_types::{
 	lp::{
 		CloseOrderJson, LimitOrRangeOrder, LimitOrder, LiquidityDepositChannelDetails,
 		OpenSwapChannels, OrderIdJson, RangeOrder, RangeOrderChange, RangeOrderSizeJson,
 		SwapRequestResponse,
 	},
-	RedemptionAmount, SwapChannelInfo,
+	OrderFills, RedemptionAmount, SwapChannelInfo,
 };
 use cf_utilities::{rpc::NumberOrHex, try_parse_number_or_hex};
 use frame_support::BoundedVec;
-use jsonrpsee::{core::async_trait, proc_macros::rpc, tokio, PendingSubscriptionSink};
+use jsonrpsee::{core::async_trait, tokio, PendingSubscriptionSink};
 use pallet_cf_ingress_egress::DepositChannelDetails;
 use pallet_cf_pools::{CloseOrder, IncreaseOrDecrease, MAX_ORDERS_DELETE};
 use sc_client_api::{
@@ -72,146 +73,6 @@ pub mod lp_crypto {
 	pub const LP_KEY_TYPE_ID: KeyTypeId = KeyTypeId(*b"lqpr");
 
 	app_crypto!(sr25519, LP_KEY_TYPE_ID);
-}
-
-#[rpc(server, client, namespace = "lp")]
-pub trait LpSignedApi {
-	#[method(name = "register_account")]
-	async fn register_account(&self) -> RpcResult<state_chain_runtime::Hash>;
-
-	#[deprecated(note = "Use `request_liquidity_deposit_address` instead")]
-	#[method(name = "liquidity_deposit")]
-	async fn request_liquidity_deposit_address_legacy(
-		&self,
-		asset: Asset,
-		wait_for: Option<WaitFor>,
-		boost_fee: Option<BasisPoints>,
-	) -> RpcResult<ApiWaitForResult<AddressString>>;
-
-	#[method(name = "request_liquidity_deposit_address")]
-	async fn request_liquidity_deposit_address(
-		&self,
-		asset: Asset,
-		wait_for: Option<WaitFor>,
-		boost_fee: Option<BasisPoints>,
-	) -> RpcResult<ApiWaitForResult<LiquidityDepositChannelDetails>>;
-
-	#[method(name = "register_liquidity_refund_address")]
-	async fn register_liquidity_refund_address(
-		&self,
-		chain: ForeignChain,
-		address: AddressString,
-	) -> RpcResult<Hash>;
-
-	#[method(name = "withdraw_asset")]
-	async fn withdraw_asset(
-		&self,
-		amount: NumberOrHex,
-		asset: Asset,
-		destination_address: AddressString,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<EgressId>>;
-
-	#[method(name = "transfer_asset")]
-	async fn transfer_asset(
-		&self,
-		amount: U256,
-		asset: Asset,
-		destination_account: AccountId32,
-	) -> RpcResult<Hash>;
-
-	#[method(name = "update_range_order")]
-	async fn update_range_order(
-		&self,
-		base_asset: Asset,
-		quote_asset: Asset,
-		id: OrderIdJson,
-		tick_range: Option<Range<Tick>>,
-		size_change: IncreaseOrDecrease<RangeOrderSizeJson>,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<Vec<RangeOrder>>>;
-
-	#[method(name = "set_range_order")]
-	async fn set_range_order(
-		&self,
-		base_asset: Asset,
-		quote_asset: Asset,
-		id: OrderIdJson,
-		tick_range: Option<Range<Tick>>,
-		size: RangeOrderSizeJson,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<Vec<RangeOrder>>>;
-
-	#[method(name = "update_limit_order")]
-	async fn update_limit_order(
-		&self,
-		base_asset: Asset,
-		quote_asset: Asset,
-		side: Side,
-		id: OrderIdJson,
-		tick: Option<Tick>,
-		amount_change: IncreaseOrDecrease<NumberOrHex>,
-		dispatch_at: Option<BlockNumber>,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<Vec<LimitOrder>>>;
-
-	#[method(name = "set_limit_order")]
-	async fn set_limit_order(
-		&self,
-		base_asset: Asset,
-		quote_asset: Asset,
-		side: Side,
-		id: OrderIdJson,
-		tick: Option<Tick>,
-		sell_amount: NumberOrHex,
-		dispatch_at: Option<BlockNumber>,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<Vec<LimitOrder>>>;
-
-	#[method(name = "free_balances", aliases = ["lp_asset_balances"])]
-	async fn free_balances(&self) -> RpcResult<AssetMap<U256>>;
-
-	#[method(name = "get_open_swap_channels")]
-	async fn get_open_swap_channels(&self, at: Option<Hash>) -> RpcResult<OpenSwapChannels>;
-
-	#[method(name = "request_redemption")]
-	async fn request_redemption(
-		&self,
-		redeem_address: EthereumAddress,
-		exact_amount: Option<NumberOrHex>,
-		executor_address: Option<EthereumAddress>,
-	) -> RpcResult<Hash>;
-
-	#[subscription(name = "subscribe_order_fills", item = BlockUpdate<OrderFills>)]
-	async fn subscribe_order_fills(&self);
-
-	#[method(name = "order_fills")]
-	async fn order_fills(&self, at: Option<Hash>) -> RpcResult<BlockUpdate<OrderFills>>;
-
-	#[method(name = "cancel_all_orders")]
-	async fn cancel_all_orders(
-		&self,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<Vec<ApiWaitForResult<Vec<LimitOrRangeOrder>>>>;
-
-	#[method(name = "cancel_orders_batch")]
-	async fn cancel_orders_batch(
-		&self,
-		orders: BoundedVec<CloseOrderJson, ConstU32<MAX_ORDERS_DELETE>>,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<Vec<LimitOrRangeOrder>>>;
-
-	#[method(name = "schedule_swap")]
-	async fn schedule_swap(
-		&self,
-		amount: NumberOrHex,
-		input_asset: Asset,
-		output_asset: Asset,
-		retry_duration: BlockNumber,
-		min_price: Price,
-		dca_params: Option<DcaParameters>,
-		wait_for: Option<WaitFor>,
-	) -> RpcResult<ApiWaitForResult<SwapRequestResponse>>;
 }
 
 /// An LP signed RPC extension for the state chain node.
@@ -276,7 +137,7 @@ where
 }
 
 #[async_trait]
-impl<C, B, BE> LpSignedApiServer for LpSignedRpc<C, B, BE>
+impl<C, B, BE> LpRpcApiServer for LpSignedRpc<C, B, BE>
 where
 	B: BlockT<Hash = state_chain_runtime::Hash, Header = state_chain_runtime::Header>,
 	BE: Send + Sync + 'static + Backend<B>,
